refactor: 컴포저블 패키지 구조화 및 `@Preview` 추가
- `BottariSearchBar`, `BottariHashChipSearchBar` 컴포저블을 `core:ui` 모듈 내의 `component/searchbar` 하위 패키지로 이동하여 파일을 재구성했습니다.
- `MainTemplateScreen`에서 변경된 `BottariHashChipSearchBar`의 import 경로를 수정했습니다.
- `BottariCheckIndicator` 컴포저블에 `@Preview`를 추가하여 개발 중 시각적인 확인이 용이하도록 개선했습니다.
