+++ android/Bottari/core/ui/src/main/java/com/bottari/core/ui/component/BottariCheckIndicator.kt
@@ -5,11 +5,14 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.bottari.bottari.designsystem.theme.BottariTheme
 
@@ -45,4 +48,19 @@ private fun generateIndicatorSize(
     if (totalQuantity <= 0) return 0F
     val safeChecked = checkedQuantity.coerceIn(0, totalQuantity)
     return safeChecked.toFloat() / totalQuantity.toFloat()
+}
+
+@Composable
+@Preview(showBackground = true)
+private fun BottariCheckIndicatorPreview() {
+    BottariTheme {
+        BottariCheckIndicator(
+            checkedQuantity = 2,
+            totalQuantity = 5,
+            modifier =
+                Modifier
+                    .size(height = 20.dp, width = 300.dp)
+                    .padding(4.dp),
+        )
+    }
 }
