+++ android/Bottari/core/ui/src/main/java/com/bottari/core/ui/component/searchbar/BottariHashChipSearchBar.kt
@@ -1,4 +1,4 @@
-package com.bottari.core.ui.component
+package com.bottari.core.ui.component.searchbar
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -30,6 +30,7 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.bottari.bottari.designsystem.theme.BottariTheme
+import com.bottari.core.ui.component.HashChipTextField
 import com.bottari.core.ui.extension.dropShadow
 
 @Composable
